restructured configuration models
